+++ src/test/java/kr/hhplus/be/server/application/point/PointCriteriaFixture.kt
@@ -2,9 +2,10 @@ package kr.hhplus.be.server.application.point
 
 import kr.hhplus.be.server.domain.auth.Authentication
 import org.mockito.Mockito.mock
+import java.math.BigDecimal
 
 class ChargeFixture(
-    amount: Long = 100,
+    amount: BigDecimal = BigDecimal.valueOf(100),
     authentication: Authentication = mock(Authentication::class.java),
 ) {
     val charge = PointCriteria.Charge(amount, authentication)
