+++ src/test/java/kr/hhplus/be/server/domain/point/PointCommandFixture.kt
@@ -2,16 +2,17 @@ package kr.hhplus.be.server.domain.point
 
 import kr.hhplus.be.server.domain.auth.Authentication
 import org.mockito.Mockito.mock
+import java.math.BigDecimal
 
 class ChargeFixture(
-    amount: Long = 100,
+    amount: BigDecimal = BigDecimal.valueOf(100),
     authentication: Authentication = mock(Authentication::class.java),
 ) {
     val charge = PointCommand.Charge(amount, authentication)
 }
 
 class UseFixture(
-    amount: Long = 100,
+    amount: BigDecimal = BigDecimal.valueOf(100),
     authentication: Authentication = mock(Authentication::class.java),
 ) {
     val charge = PointCommand.Use(amount, authentication)
